strands-demo/demo-aws-price-web/requirements.txt
flask
boto3
strands-agents
marked